Enable using `logging` module for logs (#10)
* Enable logging for driver.py

* Enable logging for db.py

* Simplify statements

* Use logging levels to modify verbosity

* Optimize imports

* Add logger

* Optimize logger usage

* Fix style

+++ db.py
@@ -1,6 +1,9 @@
-import psycopg2
-import os
 import datetime
+import os
+
+import psycopg2
+
+from logger import logger
 
 
 class DB:
@@ -13,7 +16,7 @@ class DB:
 
     __verbose = False
 
-    def __init__(self, verbose=False):
+    def __init__(self):
         self.__conn = psycopg2.connect(
             database=os.environ['POSTGRES_DB'],
             user=os.environ['POSTGRES_USER'],
@@ -22,7 +25,6 @@ class DB:
             port="5432",
         )
         self.__cursor = self.__conn.cursor()
-        self.__verbose = verbose
 
     def __del__(self):
         self.__cursor.close()
@@ -32,11 +34,10 @@ class DB:
         """Fetches the Telegram Bot API Token from the database.
         """
         try:
-            self.__cursor.execute(
-                ''' SELECT data from settings WHERE id = 2 ''')
+            self.__cursor.execute(''' SELECT data from settings WHERE id = 2 ''')
             return self.__cursor.fetchone()[0]
         except:
-            print("Telegram Bot Secret is not in the database")
+            logger.critical("Telegram Bot Secret is not in the database")
             exit()
 
     def get_offset(self):
@@ -51,14 +52,11 @@ class DB:
         """
         res = True
         try:
-            self.__cursor.execute(
-                ''' UPDATE settings SET data = (%s) WHERE id = 1 ''',
-                (offset, ))
-            if self.__verbose:
-                print("Offset updated with %s" % offset)
+            self.__cursor.execute(''' UPDATE settings SET data = (%s) WHERE id = 1 ''', (offset,))
+            logger.debug("Offset updated with %s" % offset)
         except:
             res = False
-            print("Offset could not be updated")
+            logger.warning("Offset could not be updated")
         finally:
             self.__conn.commit()
             return res
@@ -67,13 +65,9 @@ class DB:
         """Checks if a user with the given user id exists.
         Returns boolean.
         """
-        self.__cursor.execute(
-            ''' SELECT COUNT(*) from people WHERE uid = (%s) ''', (uid, ))
+        self.__cursor.execute(''' SELECT COUNT(*) from people WHERE uid = (%s) ''', (uid,))
         status = self.__cursor.fetchone()[0]
-        if status == 1:
-            return True
-        else:
-            return False
+        return status == 1
 
     def add_user(self, uid, fname, lname, uname):
         """Adds a user with the given id, first name, last name, and username to the database.
@@ -81,23 +75,16 @@ class DB:
         """
         res = True
         if self.check_if_user_exists(uid):
-            if self.__verbose:
-                print("User %s %s (%s) exists in the database" %
-                      (fname, lname, uname))
+            logger.debug("User %s %s (%s) exists in the database" % (fname, lname, uname))
             return res
         else:
             try:
-                self.__cursor.execute(
-                    ''' INSERT into people VALUES((%s),(%s),(%s),(%s)) ''',
-                    (uid, fname, lname, uname))
-                if self.__verbose:
-                    print("User %s %s (%s) created succesfully" %
-                          (fname, lname, uname))
+                self.__cursor.execute(''' INSERT into people VALUES((%s),(%s),(%s),(%s)) ''',
+                                      (uid, fname, lname, uname))
+                logger.debug("User %s %s (%s) created successfully" % (fname, lname, uname))
             except:
                 res = False
-                if self.__verbose:
-                    print("User %s %s (%s) cannot be inserted" %
-                          (fname, lname, uname))
+                logger.warning("User %s %s (%s) cannot be inserted" % (fname, lname, uname))
             finally:
                 self.__conn.commit()
                 return res
@@ -106,13 +93,9 @@ class DB:
         """Checks if a group with the given group id exists.
         Returns boolean.
         """
-        self.__cursor.execute(
-            ''' SELECT COUNT(*) from groups WHERE gid = (%s) ''', (gid, ))
+        self.__cursor.execute(''' SELECT COUNT(*) from groups WHERE gid = (%s) ''', (gid,))
         status = self.__cursor.fetchone()[0]
-        if status == 1:
-            return True
-        else:
-            return False
+        return status == 1
 
     def add_group(self, gid, title):
         """Adds a group with the given id, and title to the database.
@@ -120,33 +103,20 @@ class DB:
         """
         res = True
         if self.check_if_group_exists(gid):
-            if self.__verbose:
-                print("Group %s (%s) exists in the database" % (title, gid))
+            logger.debug("Group %s (%s) exists in the database" % (title, gid))
             return res
         else:
             try:
-                self.__cursor.execute(
-                    ''' INSERT into groups VALUES((%s),(%s)) ''', (gid, title))
-                if self.__verbose:
-                    print("Group %s (%s) created succesfully" % (title, gid))
+                self.__cursor.execute(''' INSERT into groups VALUES((%s),(%s)) ''', (gid, title))
+                logger.debug("Group %s (%s) created successfully" % (title, gid))
             except:
                 res = False
-                if self.__verbose:
-                    print("Group %s (%s) cannot be inserted" % (title, gid))
+                logger.warning("Group %s (%s) cannot be inserted" % (title, gid))
             finally:
                 self.__conn.commit()
                 return res
 
-    def log(self,
-            uid,
-            fname,
-            lname,
-            uname,
-            mid,
-            time,
-            content,
-            gid,
-            gtitle='private'):
+    def log(self, uid, fname, lname, uname, mid, time, content, gid, gtitle='private'):
         """Logs the activity to the database.
         Returns boolean.
         """
@@ -155,15 +125,12 @@ class DB:
             self.add_group(gid, gtitle)
         res = True
         try:
-            self.__cursor.execute(
-                ''' INSERT into logs VALUES((%s),(%s),(%s),(%s),(%s)) ''',
-                (uid, mid, time, content, gid))
-            if self.__verbose:
-                print("Message %s logged succesfully" % (mid))
+            self.__cursor.execute(''' INSERT into logs VALUES((%s),(%s),(%s),(%s),(%s)) ''',
+                                  (uid, mid, time, content, gid))
+            logger.debug("Message %s logged successfully" % (mid))
         except:
             res = False
-            if self.__verbose:
-                print("Message %s cannot be logged" % (mid))
+            logger.warning("Message %s cannot be logged" % (mid))
         finally:
             self.__conn.commit()
             return res
@@ -172,8 +139,7 @@ class DB:
         """Fetches the service title with the given service id from the database.
         """
         try:
-            self.__cursor.execute(
-                ''' SELECT name from services WHERE id = (%s) ''', (sid, ))
+            self.__cursor.execute(''' SELECT name from services WHERE id = (%s) ''', (sid,))
             return self.__cursor.fetchone()[0]
         except:
             return False
@@ -182,14 +148,10 @@ class DB:
         """Checks if a user with the given user id subscribed to the service.
         Returns boolean.
         """
-        self.__cursor.execute(
-            ''' SELECT COUNT(*) from subscriptions WHERE uid = (%s) and service = (%s)''',
-            (uid, service))
+        self.__cursor.execute(''' SELECT COUNT(*) from subscriptions WHERE uid = (%s) and service = (%s)''',
+                              (uid, service))
         status = self.__cursor.fetchone()[0]
-        if status == 1:
-            return True
-        else:
-            return False
+        return status == 1
 
     def add_service(self, uid, service):
         """Adds a service subscription to the database.
@@ -197,22 +159,14 @@ class DB:
         """
         res = True
         if self.check_service(uid, service) is True:
-            if self.__verbose:
-                print("User %s already subscribed to service %s" %
-                      (uid, self.get_service_title(service)))
+            logger.debug("User %s already subscribed to service %s" % (uid, self.get_service_title(service)))
             return res
         try:
-            self.__cursor.execute(
-                ''' INSERT into subscriptions VALUES((%s),(%s)) ''',
-                (uid, service))
-            if self.__verbose:
-                print("User %s subscribed to %s service succesfully" %
-                      (uid, self.get_service_title(service)))
+            self.__cursor.execute(''' INSERT into subscriptions VALUES((%s),(%s)) ''', (uid, service))
+            logger.debug("User %s subscribed to %s service successfully" % (uid, self.get_service_title(service)))
         except:
             res = False
-            if self.__verbose:
-                print("User %s cannot subscribe to %s service" %
-                      (uid, self.get_service_title(service)))
+            logger.warning("User %s cannot subscribe to %s service" % (uid, self.get_service_title(service)))
         finally:
             self.__conn.commit()
             return res
@@ -223,22 +177,14 @@ class DB:
         """
         res = False
         if self.check_service(uid, service) is False:
-            if self.__verbose:
-                print("User %s has not subscribed to service %s" %
-                      (uid, self.get_service_title(service)))
+            logger.debug("User %s has not subscribed to service %s" % (uid, self.get_service_title(service)))
             return res
         try:
-            self.__cursor.execute(
-                ''' DELETE from subscriptions WHERE uid = (%s) and service = (%s) ''',
-                (uid, service))
+            self.__cursor.execute(''' DELETE from subscriptions WHERE uid = (%s) and service = (%s) ''', (uid, service))
             res = True
-            if self.__verbose:
-                print("User %s unsubscribed from %s service succesfully" %
-                      (uid, self.get_service_title(service)))
+            logger.debug("User %s unsubscribed from %s service successfully" % (uid, self.get_service_title(service)))
         except:
-            if self.__verbose:
-                print("User %s cannot unsubscribe to %s service" %
-                      (uid, self.get_service_title(service)))
+            logger.warning("User %s cannot unsubscribe to %s service" % (uid, self.get_service_title(service)))
         finally:
             self.__conn.commit()
             return res
@@ -249,14 +195,9 @@ class DB:
         """
         now = datetime.datetime.now()
         today = str(now.year) + str(now.month) + str(now.day)
-        self.__cursor.execute(
-            ''' SELECT COUNT(*) from servicedays WHERE id = (%s) and day = (%s) ''',
-            (service, today))
+        self.__cursor.execute(''' SELECT COUNT(*) from servicedays WHERE id = (%s) and day = (%s) ''', (service, today))
         status = self.__cursor.fetchone()[0]
-        if status == 1:
-            return True
-        else:
-            return False
+        return status == 1
 
     def mark_service_sent_today(self, service):
         """Updates a service blast as sent today.
@@ -266,15 +207,11 @@ class DB:
         today = str(now.year) + str(now.month) + str(now.day)
         res = True
         try:
-            self.__cursor.execute(
-                ''' INSERT into servicedays VALUES((%s),(%s)) ''',
-                (service, today))
-            if self.__verbose:
-                print("Service days for %s updated as sent" %
-                      self.get_service_title(service))
+            self.__cursor.execute(''' INSERT into servicedays VALUES((%s),(%s)) ''', (service, today))
+            logger.debug("Service days for %s updated as sent" % self.get_service_title(service))
         except:
             res = False
-            print("Service days could not be updated")
+            logger.warning("Service days could not be updated")
         finally:
             self.__conn.commit()
             return res
@@ -283,8 +220,7 @@ class DB:
         """Fetches the group title with the given group id from the database.
         """
         if self.check_if_group_exists(gid):
-            self.__cursor.execute(
-                ''' SELECT title from groups WHERE gid = (%s) ''', (gid, ))
+            self.__cursor.execute(''' SELECT title from groups WHERE gid = (%s) ''', (gid,))
             return self.__cursor.fetchone()[0]
         else:
             return False
@@ -293,8 +229,7 @@ class DB:
         """Fetches the user details with the given user id from the database.
         """
         if self.check_if_user_exists(uid):
-            self.__cursor.execute(
-                ''' SELECT * from people WHERE uid = (%s) ''', (uid, ))
+            self.__cursor.execute(''' SELECT * from people WHERE uid = (%s) ''', (uid,))
             return self.__cursor.fetchone()
         else:
             return False
@@ -302,79 +237,69 @@ class DB:
     def get_service_users(self, service):
         """Fetches the user ids subscribed to the given service from the database.
         """
-        self.__cursor.execute(
-            ''' SELECT * from subscriptions WHERE service = (%s) ''',
-            (service, ))
+        self.__cursor.execute(''' SELECT * from subscriptions WHERE service = (%s) ''', (service,))
         return self.__cursor.fetchall()
 
     def list_logs(self):
         """Prints all logs from the database.
         """
         self.__cursor.execute(''' SELECT * FROM logs ORDER by mid ''')
-        print("Printing all logs")
+        logger.debug("Printing all logs")
         for i in self.__cursor.fetchall():
-            print(i)
+            logger.debug(i)
 
     def list_logs_from_user(self, uid):
         """Prints all logs with the given user id from the database.
         """
         if self.check_if_user_exists(uid):
             user = self.get_user(uid)
-            self.__cursor.execute(
-                ''' SELECT * FROM logs WHERE uid = (%s) ORDER by mid ''',
-                (uid, ))
-            print("Printing all logs from user %s %s (%s)" %
-                  (user[1], user[2], user[3]))
+            self.__cursor.execute(''' SELECT * FROM logs WHERE uid = (%s) ORDER by mid ''', (uid,))
+            logger.debug("Printing all logs from user %s %s (%s)" % (user[1], user[2], user[3]))
             for i in self.__cursor.fetchall():
-                print(i)
+                logger.debug(i)
         else:
-            print("No user exists with User ID %s" % uid)
+            logger.debug("No user exists with User ID %s" % uid)
 
     def list_logs_with_user(self):
         """Prints all logs and the user details as a joined table from the database.
         """
-        self.__cursor.execute(
-            ''' SELECT * FROM logs INNER JOIN people ON logs.uid = people.uid ORDER by mid '''
-        )
-        print("Printing all logs with user data")
+        self.__cursor.execute(''' SELECT * FROM logs INNER JOIN people ON logs.uid = people.uid ORDER by mid ''')
+        logger.debug("Printing all logs with user data")
         for i in self.__cursor.fetchall():
-            print(i)
+            logger.debug(i)
 
     def list_logs_from_group(self, gid):
         """Prints all logs with the given group id from the database.
         """
         if self.check_if_group_exists(gid):
             gtitle = self.get_group_title(gid)
-            self.__cursor.execute(
-                ''' SELECT * FROM logs WHERE gid = (%s) ORDER by mid ''',
-                (gid, ))
-            print("Printing all logs from group %s (%s)" % (gtitle, gid))
+            self.__cursor.execute(''' SELECT * FROM logs WHERE gid = (%s) ORDER by mid ''', (gid,))
+            logger.debug("Printing all logs from group %s (%s)" % (gtitle, gid))
             for i in self.__cursor.fetchall():
-                print(i)
+                logger.debug(i)
         else:
-            print("No group exists with Group ID %s" % gid)
+            logger.debug("No group exists with Group ID %s" % gid)
 
     def list_db(self):
         """Prints all table names from the database.
         """
-        self.__cursor.execute(
-            ''' SELECT name FROM sqlite_master WHERE type='table' ''')
-        print("Printing all databases")
+        self.__cursor.execute(''' SELECT name FROM sqlite_master WHERE type='table' ''')
+        logger.debug("Printing all databases")
         for i in self.__cursor.fetchall():
-            print(i)
+            logger.debug(i)
 
     def list_users(self):
         """Prints all users from the database.
         """
         self.__cursor.execute(''' SELECT * FROM people ORDER by uid ''')
-        print("Printing all users")
+        logger.debug("Printing all users")
         for i in self.__cursor.fetchall():
-            print(i)
+            logger.debug(i)
 
     def list_groups(self):
         """Prints all groups from the database.
         """
         self.__cursor.execute(''' SELECT * FROM groups ORDER by gid ''')
-        print("Printing all groups")
+        logger.debug("Printing all groups")
         for i in self.__cursor.fetchall():
-            print(i)
+            logger.debug(i)
